# api/openapi.yaml
openapi: 3.0.3
info:
  title: OwnDock API
  version: 0.1.0
  description: Pre-release browser, management, and automation HTTP contract. Identity, the read-only built-in Template catalog, Managed Host and Agent enrollment, Project-scoped Application, Release, Registry Credential, Source Repository connection, Build Configuration, manual, token, and signed Webhook Build trigger/query, bounded incremental Build logs, Artifact, read-only digest-bound Artifact Evidence metadata and verified content download, Release handoff, Environment, Runtime Target, Deployment lifecycle, Runtime Inventory, controlled Terminal Session, and Audit operations form the implemented product slice when product mode is enabled. Every formal Product API operation is subject to shared source and installation-wide ingress admission and can return 429 rate_limited or fail-closed 503 ingress_protection_unavailable with Retry-After. Browser access is same-origin by default; configured cross-origin clients use exact origins without REST authentication cookies, and every `/api/` response is non-cacheable. Terminal connection tickets are short-lived, one-time, path-scoped Secure HttpOnly cookies and are never returned in a URL or JSON body. The separate mTLS Agent Control v1 transport is documented outside this OpenAPI; its typed commands and Agent Runtime Gateway are internal transports. Top-level Application, Environment, and Deployment operations remain disabled engineering samples.
...
servers:
  - url: /
...
paths:
...
  /api/v1/managed-hosts:
    parameters:
      - $ref: "#/components/parameters/RequestID"
    get:
      tags: [Managed Hosts]
      operationId: listManagedHosts
      summary: List Organization-owned managed hosts
      security:
        - bearerAuth: []
      responses:
        "200":
          $ref: "#/components/responses/ManagedHostList"
        "401":
          $ref: "#/components/responses/APIError"
        "403":
          $ref: "#/components/responses/APIError"
        "500":
          $ref: "#/components/responses/APIError"
    post:
      tags: [Managed Hosts]
      operationId: createManagedHost
      summary: Register an Organization-owned managed host
      description: Only an Owner may register a host. Agent hosts remain enrolling until the separate enrollment flow creates a fixed Agent identity, then remain offline until the authenticated Agent Control flow establishes liveness.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateManagedHostRequest"
      responses:
        "201":
          $ref: "#/components/responses/ManagedHostCreated"
        "400":
          $ref: "#/components/responses/APIError"
        "401":
          $ref: "#/components/responses/APIError"
        "403":
          $ref: "#/components/responses/APIError"
        "409":
          $ref: "#/components/responses/APIError"
        "415":
          $ref: "#/components/responses/APIError"
        "422":
          $ref: "#/components/responses/APIError"
        "500":
          $ref: "#/components/responses/APIError"
...
  /api/v1/agent/enrollments:exchange:
    parameters:
      - $ref: "#/components/parameters/RequestID"
    post:
      tags: [Managed Hosts]
      operationId: exchangeAgentEnrollment
      summary: Exchange a one-time enrollment token and CSR for an Agent certificate
      description: This bootstrap endpoint does not accept a user session. The enrollment token is carried only in the JSON body and is consumed atomically. For response-loss recovery, the exact same token, CSR, instance, version, protocol and capability request may retrieve the original identity and certificate for 10 minutes; any changed request is rejected and no second identity or audit event is created.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExchangeAgentEnrollmentRequest"
      responses:
        "201":
          $ref: "#/components/responses/AgentCredentialsCreated"
        "400":
          $ref: "#/components/responses/APIError"
        "401":
          $ref: "#/components/responses/APIError"
        "415":
          $ref: "#/components/responses/APIError"
        "503":
          $ref: "#/components/responses/APIError"
        "500":
          $ref: "#/components/responses/APIError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: Opaque user session token sent explicitly in Authorization. The REST API does not authenticate user cookies.
    terminalTicket:
      type: apiKey
      in: cookie
      name: __Secure-owndock_terminal_ticket
      description: Short-lived, one-time terminal ticket scoped to the exact connect path. It is never returned in JSON or accepted in the URL.
    buildTriggerAuth:
      type: http
      scheme: bearer
      bearerFormat: OwnDock-Build-Trigger
  parameters:
    RequestID:
      name: X-Request-ID
      in: header
      required: false
      description: Optional caller-provided request correlation identifier.
      schema:
        type: string
        minLength: 1
        maxLength: 128
        pattern: "^[A-Za-z0-9._-]+$"
...
  headers:
    RequestID:
      description: Request correlation identifier accepted from or generated by OwnDock.
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 128
        pattern: "^[A-Za-z0-9._-]+$"
  responses:
...
    ManagedHostList:
      description: Organization managed host collection.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestID"
      content:
        application/json:
          schema:
            type: object
            additionalProperties: false
            required: [items]
            properties:
              items:
                type: array
                items:
                  $ref: "#/components/schemas/ManagedHost"
    ManagedHostCreated:
      description: Managed host registered.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestID"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ManagedHost"
...
    AgentCredentialsCreated:
      description: Newly issued Agent client identity and certificate chain.
      headers:
        Cache-Control:
          schema:
            type: string
        X-Request-ID:
          $ref: "#/components/headers/RequestID"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/AgentCredentials"
...
    APIError:
      description: Request failed. The safe message is localized from Accept-Language; clients branch on the stable error.code, never on message text.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestID"
        Content-Language:
          description: Language of the safe error message selected from Accept-Language. Unsupported or missing languages fall back to en-US.
          schema:
            type: string
            enum: [en-US, zh-CN]
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    ManagedHost:
      type: object
      additionalProperties: false
      required: [id, organization_id, name, status, connection_mode, capabilities, created_by, created_at, updated_at]
      properties:
        id:
          type: string
          minLength: 1
        organization_id:
          type: string
          minLength: 1
        name:
          type: string
          minLength: 2
          maxLength: 80
        status:
          type: string
          enum: [enrolling, online, offline, disabled]
        connection_mode:
          type: string
          enum: [agent, direct]
        agent_identity_id:
          type: string
          minLength: 1
          readOnly: true
        agent_instance_id:
          type: string
          minLength: 1
          maxLength: 128
          readOnly: true
        agent_certificate_expires_at:
          type: string
          format: date-time
          readOnly: true
        direct_ssh_ref:
          type: string
          pattern: "^secret://[a-z0-9][a-z0-9-]{0,62}$"
          description: External private-key reference; returned as metadata, never resolved into the API response.
        direct_ssh_address:
          type: string
          minLength: 3
          maxLength: 255
          description: Fixed SSH host and port, for example host.example.com:22.
        direct_ssh_user:
          type: string
          minLength: 1
          maxLength: 32
          pattern: "^[A-Za-z0-9_][A-Za-z0-9_.-]{0,31}$"
        direct_ssh_host_key_sha256:
          type: string
          pattern: "^SHA256:[A-Za-z0-9+/]{43}$"
          description: Pinned OpenSSH SHA-256 host-key fingerprint. Host-key verification cannot be disabled.
        last_seen_at:
          type: string
          format: date-time
        agent_version:
          type: string
          minLength: 1
        protocol_version:
          type: string
          minLength: 1
        capabilities:
          type: array
          items:
            type: string
            minLength: 1
        created_by:
          type: string
          minLength: 1
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    ExchangeAgentEnrollmentRequest:
      type: object
      additionalProperties: false
      required: [enrollment_token, instance_id, agent_version, protocol_version, capabilities, csr_pem]
      properties:
        enrollment_token:
          type: string
          minLength: 32
          writeOnly: true
        instance_id:
          type: string
          minLength: 1
          maxLength: 128
          pattern: "^[A-Za-z0-9_-][A-Za-z0-9._-]*$"
        agent_version:
          type: string
          minLength: 1
          maxLength: 64
        protocol_version:
          type: string
          minLength: 1
          maxLength: 64
        capabilities:
          description: Capabilities approved for this Agent Identity. A later authenticated hello may advertise only a subset of this grant.
          type: array
          maxItems: 16
          uniqueItems: true
          items:
            type: string
            minLength: 1
            maxLength: 64
            pattern: "^[a-z0-9][a-z0-9._-]{0,63}$"
        csr_pem:
          type: string
          minLength: 1
          maxLength: 16384
          writeOnly: true
    AgentCredentials:
      type: object
      additionalProperties: false
      required: [agent_identity_id, managed_host_id, certificate_pem, ca_certificate_pem, certificate_expires_at]
      properties:
        agent_identity_id:
          type: string
          minLength: 1
        managed_host_id:
          type: string
          minLength: 1
        certificate_pem:
          type: string
          minLength: 1
          readOnly: true
        ca_certificate_pem:
          type: string
          minLength: 1
          readOnly: true
        certificate_expires_at:
          type: string
          format: date-time
...
    CreateManagedHostRequest:
      oneOf:
        - $ref: "#/components/schemas/CreateAgentManagedHostRequest"
        - $ref: "#/components/schemas/CreateDirectManagedHostRequest"
    CreateAgentManagedHostRequest:
      type: object
      additionalProperties: false
      required: [name, connection_mode]
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 80
        connection_mode:
          type: string
          enum: [agent]
    CreateDirectManagedHostRequest:
      type: object
      additionalProperties: false
      required: [name, connection_mode]
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 80
        connection_mode:
          type: string
          enum: [direct]
        direct_ssh_ref:
          type: string
          pattern: "^secret://[a-z0-9][a-z0-9-]{0,62}$"
          description: Optional direct SSH private-key reference. If set, address, user and host-key fingerprint are all required.
        direct_ssh_address:
          type: string
          minLength: 3
          maxLength: 255
          description: Fixed SSH host and port, for example host.example.com:22.
        direct_ssh_user:
          type: string
          minLength: 1
          maxLength: 32
          pattern: "^[A-Za-z0-9_][A-Za-z0-9_.-]{0,31}$"
        direct_ssh_host_key_sha256:
          type: string
          pattern: "^SHA256:[A-Za-z0-9+/]{43}$"
...
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/ErrorDetail"
    ErrorDetail:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code:
          type: string
          minLength: 1
        message:
          type: string
          minLength: 1
        request_id:
          type: string
          minLength: 1